% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_title}
\alias{generate_title}
\title{Generate title (h1 level) in \href{../../flexdashboard/html/00Index.html}{flexdashboard} style}
\usage{
generate_title(title, nav)
}
\arguments{
\item{title}{the content of title.}

\item{nav}{which (what) menu the title belong to.}
}
\value{
Pretty printed (by \link{cat} but not \link{print}) title.
}
\description{
Generate title (h1 level) in \href{../../flexdashboard/html/00Index.html}{flexdashboard} style
}
\examples{
generate_title("Qi Fat and Yu Fat", "triple-six")
}
