% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LncPipeReporter.R
\docType{package}
\name{LncPipeReporter}
\alias{LncPipeReporter}
\alias{LncPipeReporter-package}
\title{LncPipeReporter: Automatically Aggregating and Summarizing Bioinformatics Results for Interactive Report.}
\description{
The LncPipeReporter package provides three categories of important functions:
foo, bar and baz.
}
\section{LncPipeReporter functions}{

The foo functions ...
}

