% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_block_data}
\alias{test_block_data}
\title{Test and print if a specific kind of file exists.}
\usage{
test_block_data(l)
}
\arguments{
\item{l}{list of file type. Generally produced by \link{search_then_determine}.}
}
\value{
None
}
\description{
Test and print if a specific kind of file exists.
}
