% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toUpperFirstLetter}
\alias{toUpperFirstLetter}
\title{Change the first letter of sentence/phrase to uppercase.}
\usage{
toUpperFirstLetter(x)
}
\arguments{
\item{x}{the sentence/phrase.}
}
\value{
The sentence/phrase converted.
}
\description{
Change the first letter of sentence/phrase to uppercase.
}
